re-establishes v1 login connection test successful
Reenables the v1 connection test. Now v1 and v2 connection test work.
